+++ R/constructor.R
@@ -75,7 +75,7 @@ new_function <- function(args, body, env) {
   f
 }
 missing_args <- function(names) {
-  lapply(setNames(, names), function(i) quote(missing_class))
+  lapply(setNames(, names), function(i) quote(class_missing))
 }
 new_call <- function(call, args) {
   as.call(c(list(as.name(call)), args))