+++ R/convert.R
@@ -15,21 +15,21 @@
 #' @param ... Other arguments passed to custom `convert()` methods.
 #' @export
 #' @examples
-#' foo1 <- new_class("foo1", properties = list(x = integer))
-#' foo2 <- new_class("foo2", foo1, properties = list(y = double))
+#' foo1 <- new_class("foo1", properties = list(x = class_integer))
+#' foo2 <- new_class("foo2", foo1, properties = list(y = class_double))
 #'
-#' method(convert, list(foo1, integer)) <- function(from, to) from@x
-#' method(convert, list(foo2, double)) <- function(from, to) from@y
+#' method(convert, list(foo1, class_integer)) <- function(from, to) from@x
+#' method(convert, list(foo2, class_double)) <- function(from, to) from@y
 #'
-#' convert(foo1(x = 1L), to = integer)
-#' try(convert(foo1(x = 1L), to = double))
+#' convert(foo1(x = 1L), to = class_integer)
+#' try(convert(foo1(x = 1L), to = class_double))
 #'
-#' convert(foo2(x = 1L, y = 2), to = integer)
-#' convert(foo2(x = 1L, y = 2), to = double)
+#' convert(foo2(x = 1L, y = 2), to = class_integer)
+#' convert(foo2(x = 1L, y = 2), to = class_double)
 #' convert(foo2(x = 1L, y = 2), to = foo1)
 #'
 #' # If we define a convert method for interger + foo1:
-#' method(convert, list(integer, foo1)) <- function(from, to) foo1(x = from)
+#' method(convert, list(class_integer, foo1)) <- function(from, to) foo1(x = from)
 #' convert(1L, to = foo1)
 #' # Converting too foo2 still errors
 #' try(convert(1L, to = foo2))