+++ R/generic.R
@@ -38,9 +38,9 @@
 #' @examples
 #' # A simple generic with methods for some base types and S3 classes
 #' type_of <- new_generic("type_of", dispatch_args = "x")
-#' method(type_of, "character") <- function(x, ...) "A character vector"
+#' method(type_of, class_character) <- function(x, ...) "A character vector"
 #' method(type_of, new_S3_class("data.frame")) <- function(x, ...) "A data frame"
-#' method(type_of, "function") <- function(x, ...) "A function"
+#' method(type_of, class_function) <- function(x, ...) "A function"
 #'
 #' type_of(mtcars)
 #' type_of(letters)
@@ -52,7 +52,7 @@
 #'    R7_dispatch()
 #' })
 #'
-#' method(mean2, "numeric") <- function(x, ..., na.rm = FALSE) {
+#' method(mean2, class_numeric) <- function(x, ..., na.rm = FALSE) {
 #'   if (na.rm) {
 #'     x <- x[!is.na(x)]
 #'   }
@@ -60,7 +60,7 @@
 #' }
 #'
 #' # You'll be warned if you forget the argument:
-#' method(mean2, "character") <- function(x, ...) {
+#' method(mean2, class_character) <- function(x, ...) {
 #'   stop("Not supported")
 #' }
 new_generic <- function(name, dispatch_args, fun = NULL) {