+++ R/method-introspect.R
@@ -9,16 +9,15 @@
 #' @seealso [method_explain()] to explain why a specific method was picked.
 #' @inheritParams method<-
 #' @returns A function with class <R7_method>.
-#' @param class,object Perform introspection either with `classes`
-#'   (processed with [as_class()]) or a concrete objects.
-#'
-#'   If `generic` does multiple dispatch both `object` and `class` need
-#'   to be wrapped in a list.
+#' @param class,object Perform introspection either with a `class`
+#'   (processed with [as_class()]) or a concrete `object`. If `generic` uses
+#'   multiple dispatch then both `object` and `class` must be a list of
+#'   classes/objects.
 #' @export
 #' @examples
 #' # Create a generic and register some methods
 #' bizarro <- new_generic("bizarro", "x")
-#' method(bizarro, "numeric") <- function(x) rev(x)
+#' method(bizarro, class_numeric) <- function(x) rev(x)
 #' method(bizarro, new_S3_class("factor")) <- function(x) {
 #'   levels(x) <- rev(levels(x))
 #'   x
@@ -28,7 +27,7 @@
 #' bizarro
 #'
 #' # And you can use method() to inspect specific implementations
-#' method(bizarro, class = "integer")
+#' method(bizarro, class = class_integer)
 #' method(bizarro, object = 1)
 #' method(bizarro, new_S3_class("factor"))
 method <- function(generic, class = NULL, object = NULL) {