+++ tests/testthat/test-constructor.R
@@ -5,7 +5,7 @@ test_that("generates correct arguments from parent + properties",  {
   expect_equal(args$parent, character())
 
   # Includes properties
-  args <- constructor_args(R7_object, as_properties(list(x = "numeric")))
+  args <- constructor_args(R7_object, as_properties(list(x = class_numeric)))
   expect_equal(args$self, "x")
   expect_equal(args$parent, character())
 
@@ -17,17 +17,17 @@ test_that("generates correct arguments from parent + properties",  {
   expect_equal(args$parent, character())
 
   # Includes parent properties
-  foo <- new_class("foo", properties = list(x = "numeric"))
-  args <- constructor_args(foo, as_properties(list(y = "numeric")))
+  foo <- new_class("foo", properties = list(x = class_numeric))
+  args <- constructor_args(foo, as_properties(list(y = class_numeric)))
   expect_equal(args$self, "y")
   expect_equal(args$parent, "x")
 
   # But only those in the constructor
   foo <- new_class("foo",
-    properties = list(x = "numeric"),
+    properties = list(x = class_numeric),
     constructor = function() new_object(x = 1)
   )
-  args <- constructor_args(foo, as_properties(list(y = "numeric")))
+  args <- constructor_args(foo, as_properties(list(y = class_numeric)))
   expect_equal(args$self, "y")
   expect_equal(args$parent, character())
 })
@@ -35,9 +35,9 @@ test_that("generates correct arguments from parent + properties",  {
 test_that("generates meaningful constructors", {
   expect_snapshot({
     new_constructor(R7_object, list())
-    new_constructor(R7_object, as_properties(list(x = "numeric", y = "numeric")))
+    new_constructor(R7_object, as_properties(list(x = class_numeric, y = class_numeric)))
 
-    foo <- new_class("foo", parent = "character")
+    foo <- new_class("foo", parent = class_character)
     new_constructor(foo, list())
 
     foo2 <- new_class("foo2", parent = foo)
@@ -48,14 +48,14 @@ test_that("generates meaningful constructors", {
 test_that("can generate constructors for S3 classes", {
   expect_snapshot({
     new_constructor(S3_factor, list())
-    new_constructor(S3_factor, as_properties(list(x = "numeric", y = "numeric")))
+    new_constructor(S3_factor, as_properties(list(x = class_numeric, y = class_numeric)))
   }, transform = scrub_environment)
 })
 
 test_that("can generate constructor for inherited abstract classes", {
   expect_snapshot({
-    foo1 <- new_class("foo1", abstract = TRUE, properties = list(x = double))
+    foo1 <- new_class("foo1", abstract = TRUE, properties = list(x = class_double))
     new_constructor(foo1, list())
-    new_constructor(foo1, as_properties(list(y = double)))
+    new_constructor(foo1, as_properties(list(y = class_double)))
   }, transform = scrub_environment)
 })