+++ tests/testthat/test-data.R
@@ -1,11 +1,11 @@
 describe("R7_data", {
-  text <- new_class("text", character)
+  text <- new_class("text", class_character)
   it("retrieves .data", {
     x <- text("hi")
     expect_equal(R7_data(x), "hi")
   })
   it("strips properties", {
-    text <- new_class("text", character, properties = list(x = "integer"))
+    text <- new_class("text", class_character, properties = list(x = class_integer))
     x <- text("hi", x = 10L)
     expect_equal(attributes(R7_data(x)), NULL)
   })