+++ tests/testthat/test-special.R
@@ -1,13 +1,13 @@
 test_that("can test and print", {
-  expect_true(is_missing_class(missing_class))
-  expect_false(is_missing_class(any_class))
-  expect_true(is_any_class(any_class))
-  expect_false(is_any_class(missing_class))
+  expect_true(is_class_missing(class_missing))
+  expect_false(is_class_missing(class_any))
+  expect_true(is_class_any(class_any))
+  expect_false(is_class_any(class_missing))
 
   expect_snapshot({
-    print(missing_class)
-    print(any_class)
+    print(class_missing)
+    print(class_any)
 
-    str(list(m = missing_class, a = any_class))
+    str(list(m = class_missing, a = class_any))
   })
 })